# docs/swagger.yaml
definitions:
  dtos.AllowAccessTokenDTO:
    properties:
      api_id:
        minLength: 2
        type: string
      client_id:
        minLength: 2
        type: string
      resource_method:
        minLength: 2
        type: string
      resource_path:
        minLength: 2
        type: string
      token:
        minLength: 2
        type: string
    required:
    - api_id
    - client_id
    - resource_method
    - resource_path
    - token
    type: object
  dtos.AllowAccessTokenResponseDTO:
    properties:
      message:
        type: string
    type: object
  dtos.CreateClientDTO:
    properties:
      client_name:
        minLength: 2
        type: string
      key_period:
        minLength: 2
        type: string
      key_time_duration:
        minimum: 1
        type: integer
      permissions:
        items:
          $ref: '#/definitions/dtos.Permission'
        type: array
      scope_name:
        minLength: 2
        type: string
    required:
    - client_name
    - key_period
    - key_time_duration
    - scope_name
    type: object
  dtos.CreateClientResponseDTO:
    properties:
      api_id:
        type: string
      client_name:
        type: string
      id:
        type: string
      key_period:
        type: string
      key_time_duration:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/dtos.Permission'
        type: array
      scope_id:
        type: string
      scope_name:
        type: string
    type: object
  dtos.GenerateTokenDTO:
    properties:
      api_id:
        minLength: 2
        type: string
      client_id:
        minLength: 2
        type: string
    required:
    - api_id
    - client_id
    type: object
  dtos.GenerateTokenResponseDTO:
    properties:
      token:
        type: string
    type: object
  dtos.Permission:
    properties:
      resource_method:
        type: string
      resource_name:
        type: string
      resource_path:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: http://localhost:8088
info:
  contact:
    email: mlaguardia@gmail.com
    name: Michel La Guardia
    url: https://www.github.com/castmetal
  description: This is a sample server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Token Example API
  version: "1.0"
paths:
  /v1/client:
    post:
      consumes:
      - application/json
      description: Creating a client with route permissions
      parameters:
      - description: CreateClient Data
        in: body
        name: createClient
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateClientDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CreateClientResponseDTO'
      summary: Create a client with CreateClient Data input
      tags:
      - Clients
  /v1/token/access:
    post:
      consumes:
      - application/json
      description: Allow access to a specific route
      parameters:
      - description: Generate Token Data
        in: body
        name: allowToken
        required: true
        schema:
          $ref: '#/definitions/dtos.AllowAccessTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.AllowAccessTokenResponseDTO'
      summary: Concede access to a route
      tags:
      - Token
  /v1/token/generate:
    post:
      consumes:
      - application/json
      description: Generates a new JWT token to access and authorize requesting for
        routes
      parameters:
      - description: Generate Token Data
        in: body
        name: generateToken
        required: true
        schema:
          $ref: '#/definitions/dtos.GenerateTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GenerateTokenResponseDTO'
      summary: Generate a JWT token with Generate Token input
      tags:
      - Token
swagger: "2.0"